Clarify 'GitLab' term in signup error message
When `allow_single_sign_on` option is `true`, new users can't signup in
using the SSO button without first having an account the current
instance. They need first to create an account in the current instance
and then link the account with the external provider.

For self-managed instances, we would like to clarify the difference
between `GitLab.com` account and `GitLab` account (this one refers to
the current instance) in the two sentences.

Changelog: changed