Support backend pagination for the access tokens
Personal, group and project access tokens supports backend pagination.
This is enabled using the `access_token_pagination` feature flag.
Impersonation tokens don't support backend pagination.

In this commit, we modify the frontend to support the backend
pagination. The paginator appears when the number of tokens is larger
than 20 (not configurable at the moment).

Previously, the pagination was done all in the frontend, which mean that
all the data was sent to the frontend at once. The paginator appeared
when the number of tokens were larger than 100.