Password modal to confirm two-factor actions
We created a new Vue component, TwoFactorAction, to replace
ManageTwoFactorForm and to be used to destroy the one-time password
authenticator.

The simpler TwoFactorAction component uses a better password field that
toggle the visibility and takes less space because the password is
inside the modal. By using the same modal in the whole page it makes the
experience more consistent.

The new component is under the `two_factor_actions` feature flag.

On a follow-up, we will use this new confirm modal to delete WebAuthn
device.