Document membership security for user mapping
Within the way GitLab permissions work it's illegal for a child
permission to be less than an inherited permission, so we skip creating
it.

https://gitlab.com/gitlab-org/gitlab/-/blob/65521dc8de2d01d20ffd669e7a79d516c71a919a/app/services/import/reassign_placeholder_user_records_service.rb#L127-132

Group owners need to understand clearly what can happen during import to
memberships due to inheritance. This change adds this information
to the user mapping documentation.