Post bot message for fail-open violations
This MR changes logic around fail-open policies to persist violations
for these cases so that we can highlight skipped policies in
the policy bot message.

The violations for fail-open policies are
created with a different status so they can be treated differently than
errors and so that these violations don't block the MR with
policy mergability check.

Changelog: changed
EE: true