Use dynamic primary key helpers for querying instead of ids
Remove an unnecessary spec
The value being tested here is an enum attribute of PipelineMetadata.
There's no particularly plausible way for an "invalid" attribute to
appear here. Even if one were to appear, the behavior of this service
would just be to do nothing, and not cancel the currently running
pipeline.