Fix null byte error when X-ray saves reports
This MR fixes a JSON error that occurs when
Projects::XrayReport.upsert_all is executed with
a value that contains a unicode null byte.

Null bytes are not permitted by JSON and will
also raise an error if you attempt to insert it
into a text-based PostgreSQL column.

This MR fixes this issue by sanitizing the file
content before we parse them.