Add db_config_database into marginalia comments
This commit adds `db_config_database` into the marginalia comments to
indicate the database used by the connection for an SQL query.

This change affects users connecting to the database through a
pgbouncer. The `db_config_database` in the marginalia comment will map
to the pgbouncer pool name. This way, different gitlab instances connecting
through separate pgbouncer pools on the same pgbouncer instance can be
distinguished.

This change does not affect users connecting directly to the database.

Changelog: other