Block MRs which introduce failing Vue 3 specs
Merge requests will be blocked when they:

- introduce new specs which fail under Vue 3;
- cause a spec that passes on `master` to fail under Vue 3;
- rename a spec that was previously failing under Vue 3.

The third scenario is the only one that I think at this stage justifies
*adding* to the local quarantine list, since it didn't cause the
original failure.

Merge requests will *not* be blocked when they modify a spec that is
already failing on `master` under Vue 3.

There will be a window where in-flight merge requests which introduce
broken specs merge around the same time as this one. Because the local
list of quarantined specs is based on the most recent nightly
[pipeline][1], it could therefore be out of date.

This is why there's a separate [project][2] for quarantining tests. I do
not expect that project to need to exist for long. A milestone is
probably more than enough time.

[1]: https://gitlab-org.gitlab.io/frontend/playground/jest-speed-reporter/vue3.html
[2]: https://gitlab.com/gitlab-org/frontend/playground/fast-jest-vue-3-quarantine