Clarify Jira config when password auth disabled
Documenting From `@anton`:

It's possible to use an enterprise user with password auth disabled on
the group to setup the GitLab for Jira Cloud app.

The only caveat is you must already be signed in as the enterprise user
on GitLab.com before configuring the GitLab for Jira Cloud app.

The reason being:

- Enterprise users can only sign in via the group's SSO page
- The GitLab for Jira Cloud app only loads the instance level sign in
  page during configuration. Enterprise users cant sign in via that page
  since password auth is disabled
- If the enterprise user (or any user for that matter) is signed in
  before the GitLab for Jira Cloud app is configured, then the app will
  use that GitLab user for the setup process