Change custom fields feature to use correct ids
The ids in work_item_types are being changed to the values in
work_item_types.correct_id.

For this new table, we now use work_item_types.correct_id so that we do
not need to change the values inserted here.

This is a new empty table so this should be safe to change.