Fix required approvals for unenforceable rules with policy check
When `policy_mergability_check` FF is enabled, the required approvals
are set to 0 after a commit because the expectation is that policy
violation will reset these approvals.

If there's no CI configuration in the project,
`UnenforceablePolicyRulesNotificationService` is triggering the bot
comment but it's not resetting the approvals back.

It is also not scheduled for consecutive commits after MR is created,
so the required approvals become optional in that case.