Add customize jira issue setting
This adds a new setting called `customize_jira_issue_enabled` to the
jira integration settings. It is disabled by default. Once enabled,
it allows creating Jira issues for vulnerabilities via the legacy way.
This redirects the user to the Jira issue form with vulnerability data
pre-filled instead of a GraphQL mutation which automatically creates
the issue.

Changelog: added
EE: true