Fix capybara wait in blobs/edit_spec
From our
[screenshots in CI](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/169071#note_2158785549)
we were able to determine that there was no diff to preview in this
spec. After looking more closely into the code we can see that the test
is relying on this `edit_and_commit` to update the code each time. But
depending on the test order we can see logically that the content is not
necessarily changed by calling `edit_and_commit`. This is because we use
`let_it_be(:project` so the project is re-used between specs. If the
last spec also calls this method then the content will be the same and
there will not be any diff to preview or commit.

We can see the spec fails consistently (without this change) if the
order is:

```
Editing file blob
  as a developer
    updates the content of file with a number as file path
    editing a template file in a sub directory does not change path
    updating file path updates syntax highlighting
    from MR diff
      returns me to the mr
    blob edit toolbar
      has defined set of toolbar buttons when the flag is on
      has defined set of toolbar buttons when the flag is off
    from blob file path
      updates content
      previews content
```

This MR fixes the problem by just appending `object_id` to the content
so that it is different every time it runs. The `object_id` is unique to
every test example.

In addition to this the previous implementation does not seem to robustly wait for the
element to appear. It does rely on `wait_for_requests` but I don't think
this is as reliable as using Capybara's default when using `have_`
matchers. The `have_css` matcher should wait for a specific amount of
time for the element to appear. And it will timeout if the element does
not appear in time. This seems more robust than counting the lines which
appear after a `wait_for_requests`. The counting was also not needed as
we're only checking for `> 0` so we may as well just wait for something
to appear.

This is in an effort to fix the flaky spec described at
https://gitlab.com/gitlab-org/gitlab/-/issues/448293 .