Use string values for Sidekiq workers
Sidekiq serializes worker args using JSON. JSON doesn't not have a
symbol notation so these keys are being converted from symbols to
strings which means what we're passing in, isn't what we're using during
the workers.

In this case we are defensively programming to convert `action.to_s` so
when we use `perform_now` or `perform_async` we're always working with
strings so we don't have much risk here but sidekiq is logging warnings
which makes the rspec output messy.

This change converts the values to strings and also removes the `to_s`
conversion that isn't necessary anymore (we should allow this to raise
if someone passes a symbol to `perform_now`)