Stop creating keep-arounds on merge base commits
Keep-arounds work by ensuring a commit is referred by a reference and so it
will not be considered orphaned. The commit that the keep-around
references also references the commits that it was based off. So by
extension those commits will also be kept. Since the head commit is kept
and since merge base commits are by definition upstream of the head
commit, this means that adding a keep-around on a merge base commit is
always superfluous.

Changelog: performance