Fix code_flow_steps_section_spec.js under Vue 3
This removes code and a test that probably should have been removed as
part of https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161687,
which removed the description element that was being passed to
`renderGFM`.

The test was passing under Vue 2 but not under Vue 3 because
`@vue/test-utils@2` throws on attempts to access `element` on an empty
wrapper, whereas `@vue/test-utils@1` only returns `undefined`.

See also
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/169185#note_2167844196.