Sync arg names between BE and FE
When a FeatureSetting is disabled, the :ai_self_hosted_model_id needs to be cleared from the record so that the self-hosted model that is potentially set before it is not associated any longer. A bug was introduced where if a self_hosted_model_id is nil, then it is removed from the resolver params and not passed on to the update service, meaning the field doesn't get updated.
This commit ensures that the self-hosted model id is always passed to the service even when nil, as intended. This commit also syncs the naming of the backend and frontend arguments so it no longer needs to be renamed in the resolver to match the backend expectation.