Optimize kubernetes_variables on pipeline creation
We call some methods for `kubernetes_variables` per job
when creating pipelines. We don't need to call for each job, most of the
time, we only need one call based on `environment` and
`kubernetes_namespace`. When creating a pipeline, `token` of a job is
`nil` so we can ignore it safely.

In this change, we are memoizing some methods and reducing the number of
calls of those methods.

These changes are behind the FF
`ci_variables_optimize_kubernetes_variables`.