Add tests to cover artifact redirect
During another change I noticed that we do not test all of the
redirects. These are pretty easy to test and don't cost much so I've
added a few specs here.

Also removed the RendersBlob module inclusion as it's no longer needed
now that we've moved the rendering logic to the project/build_artifacts
controller.