Add epic_issue and epic_links work item references
To ensure in the future that every `epic_issue` has a
`work_item_parent_link` and every `related_epic_link` has a
RelatedWorkItemLink (which is stored in the `issue_links` table) we add
foreign keys to the legacy epic tables.

We also want to ensure that all deletions on the work item side are
reflected on the legacy epic side, we add ON DELETE CASCADE to these
foreign keys.

Once we backfilled the tables and set the correct foreign key on
creation, we can add a NOT NULL constraint to harden the relationship
between legacy epics and new epic work items.

Changelog: added
EE: true