Refactor check_access method in protected ref modules
This change refactors the protected ref modules as a number of people
have noted the logic is hard to understand due to the nature of using
super and blocks to extend the behavior of check_access.

During a review of another MR it was clear that we could simplify this
by using meta programming to send the check to the correct type.

I've also extended this concept to the humanize method which now directs
to the type e.g. humanize_deploy_key

I've added comments to these methods to make them searchable.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/468688