Remove use of `missing` for ProtectedRef::for_role
`ActiveRecord::missing` uses `LEFT OUTER JOIN` to check for missing
associations.

In this case we are checking for `belongs_to` associations so we already
have the foreign keys in the table being queried. It is much more
performant to use the foreign key fields instead.

This also removes the need to mark this query as a cross table join.