Make it clearer that `scope_to` needs an index to work well
This discussion came up
[in Slack (internal)](https://gitlab.slack.com/archives/C3NBYFJ6N/p1728287295972459)
and at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166812#note_2145000488
as it seems like these docs don't highlight the risks of using `where`
clauses or `scope_to` before the `each_batch` methods. Right now we only
have a `NOTE` at the end but if someone is just quickly looking for code
snippets or looking at the headings they might miss this. Given that
many of us agreed it is usually better not to do this I opted to
highlight the importance of the index at:

1. The heading
2. The first paragraph
3. As a comment inside the code

Hopefully this should make it harder to miss