Store file digests in the Maven virtual registry
Do not cache them as regular files. Instead, store them along with the related file.
When a digest is requested, locate the related file and read the corresponding digest.

This also implements FIPS guards.
Md5 digests are not stored if GitLab is running in a FIPS environment.
Interacting with the md5 digest, in a FIPS environment will fail.