Create the `HelpPagePath/EnsureValidLinks` cop
This creates a cop that ensures that `help_page_path` usages link to
existing documentation.
As the documentation gets updated, it's easy to forget that we link to
it from many parts of the product. Those links can get outdated as
documented are removed, section moved around, titles changed, etc.
This cop should ensure the product links stay in sync with the
documentation.