E2E: Always expand breadcrumbs prior to checking for text
In some situations the breadcrumbs may be condensed if the full path
cannot be rendered in a limited amount of space. Certain tests execute
checks against text in the breadcrumbs and so in these situations if
the part of the component that is expected is hidden the test would
fail. It should be safe to expand the breadcrumb prior to checking for
text, which should still result in the the check has_breadcrumb?(text)
returnign the same true/false value.

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/497758