Refactor dropdownToggleText()
Refactor dropdownToggleText so that "Select a self-hosted model" is returned if the provider is vendored or the model can't be found. Not being able to find a model when the provider is self-hosted is buggy behaviour and needs to be fixed at the root, however this should make the component more resilient at least.

https://gitlab.com/gitlab-org/gitlab/-/issues/494660#note_2153559886