Allow changing synced work item epics
This removes the `lock_work_item_epics` feature flag, which was
disabled by default.
It allowed us to prevent any changes to the legacy epic from the work
item side. Now that work item epics are about to be rollout out we can
remove this policy check as we want to write from the work item side to
the legacy epic.

Changelog: changed
EE: true