Avoid SELECT 1/LIMIT 1 query in credentials inventory
When the query hasn't yet loaded and the view does something like
credentials.empty?, ActiveRecord executes a SELECT 1/LIMIT 1
query. If that empty call returns false then the view loads up
all the credentials to display them. This query is not performant
so we explcitly call load in controller actions to avoid this
query.

EE: true
Changelog: fixed