Attempt to make the security implementation issue docs clearer
This started in a discussion at
https://gitlab.com/gitlab-org/security/gitlab/-/merge_requests/4498#note_2138891381
where it seems like developers are often forgetting to create the
security implementation issue. The wording was still a little confusing
because it tells you to mark the issue related as the first step even
before the link to the documentation that explains how to actually
create that issue.