E2E: Update duo chat test to dynamically check mock/real backend
Adds logic to check the direct_access['base_url'] value which should
indicate where the AI Gateway is running. As we sometimes run with a
mocked backend, we can use this value to determine whether the test
shoudl expect to run a mocked string or something generated by a live
model.