Fix scripts/rspec_bisect_flaky on Mac
Currently this script is not working on Mac because `--bisect` is
forking processes before requiring `config/initializers/macos.rb`. This
initializer was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/122115 to
workaround some bug with `fork`. Without requiring this file it causes
the fork to fail with the following error:

```
Running suite to find failures...objc[10877]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called.
objc[10877]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called. We cannot safely call it or ignore it in the fork() child process. Crashing instead. Set a breakpoint on objc_initializeAfterForkError to debug.
```

This MR fixes 2 problems:

1. Update the `rspec_bisect_flaky` to explicitly call
   `--require config/initializers/macos.rb` to ensure this is required
   before bisect forks
1. The `config/initializers/macos.rb` needs an extra `require 'ethon'`
   to be run before everything else

You can see this
[Slack discussion (internal)](https://gitlab.slack.com/archives/C2Z9A056E/p1727723530427539)
for more information.