Improve docs for diagnosing single spec ordering issues
I recently updated this in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/169672 but I
misunderstood that the `rspec_check_order_dependence` is only designed
to see if specs that are marked as running in a static order are
now safe to run in any order. I also missed that we had docs already on
this page about how to diagnose ordering flakiness in a single spec
file.

This undos my changes in that MR and adds a link to the other section on
this page to improve discoverability. In addition this also adds the
`--require ./config/initializers/macos.rb` which is needed to fix an
issue on Mac OS when running `--bisect`. This problem and solution is
also described in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/170190 where we
fixed `scripts/rspec_bisect_flaky`.