Fix false positives in outdated todo danger rule due to FOSS/EE files
Prior this commit, when a file was moved (and changed) from FOSS to EE
and the TODOs where adjusted by prepending `ee/` the danger rule failed
because it saw that a FOSS file removed but still matched a substring of
the filename in TODOs.

For example:

 # Rename ./spec/services/reset_project_cache_service_spec.rb to
 # ./ee/spec/services/reset_project_cache_service_spec.rb
 #   and change the code is git sees a removal and addition.
 # Adjust spec/support/rspec_order_todo.yml and prefix ee/
 # Adjust .rubocop_todo/rspec/named_subject.yml and prefix ee/

 FROM_LEFRHOOK=1 bundle exec rake danger_local

  Errors:
  - [ ] `spec/services/reset_project_cache_service_spec.rb` was removed but is mentioned in:
  - `.rubocop_todo/rspec/named_subject.yml:3237`
  - `spec/support/rspec_order_todo.yml:7718`

After this commit danger no longer fails because of the added negative
lookbehind to ensure that filenames are not prefixed with `ee/`.