Track duration spent in concurrency limit queue
This MR updates `.concurrency_limit_resume` to set the timestamp when
the job is added to the concurrency limit queue.

We use `buffered_at` to minimise confusion between various queues.

This updates the scheduling latency to track the duration starting from
when the job is added to the concurrency limit queue.