Remove Danger warning for finalizing migration required stop
Since https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165146 we
now have a CI job that enforces a required stop has passed before we can
finalize a background migration.

Additionally these danger warnings are quite noisy. For example all the
discussions I had to resolve in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/169518 .

This note does also mention "Check the timings" and theoretically there
may be a case to wait for 2 required stops for any migration that is
very slow but I don't believe we do this in practice. For now I think
we've settled on the following requirements before finalizing:

1. It is finished on GitLab.com
2. We've passed a required stop

And we already have adequate enforcement for these so I don't think this
warning is really helping us anymore.