Increase LSIF artifacts in code
We are increasing LSIF artifacts limit so we can accommodate projects
with bigger LSIF indexes.

This is behind a feature flag and done in code for now as we want
to monitor how will this affect workhorse memory usage.

Once confirmed that it's good, we'll remove the feature flag and
add a migration to increase the default LSIF artifacts limit.