Retry RepositoryUpdateMirror jobs on failure
**Problem**

Due to a Sidekiq outage, the pull mirroring process got stuck because
RepositoryUpdateMirrorWorker jobs failed. The StuckImportJob runs a 24H
schedule and this meant the jobs were never retried after the Sidekiq
was brought back online.

**Solution**

Introduce retries into RepositoryUpdateMirrorWorker to serve as a first
layer recovery for situations like this in the future

Changelog: fixed
EE: true