apps/settings: Extend maximum token and SSH key expiration to 400 days
An expiration of 365 days means that rotation ends up "creeping" back
the calendar every year unless tokens are rotated on the exact day they
expire. A forced "lose a day" occurs when the validity period spans over
February 29th. In practice, the expiration falling on a weekend would
also force losing a day or two. Instead, bump the maximum allowed to 400
days for all editions so that a little over one month of leeway is
allowed to perform rotations of tokens and SSH keys.

While 395 would be "sufficient", 400 being so close makes it worth
choosing instead.

Changelog: changed