Move base integration code to a separate module
In order to support migration of instance integrations
out of integrations table we need to create a new instance
integration model. In order to do that, we need to extract
existing integration behavior from Integration model
and put it to a module, so it can be used in both
instance/non-instance integration models.

Changelog: other