Add deduplication configuration to AutoMergeProcessWorker
The safer-seeming option here would be to duplicate until the first
job is fully finished executing, to avoid a race condition with two
simultaneous jobs. Rescheduling once allows the job that hits the
queue last to try again in a moment, ensuring that we do merge the MR
but that the second will more gracfully exit if the MR is merged by
the first job.