Change CI/CD settings link in run pipeline
The CI/CD Settings page is only accessible to maintainers or above. But
the link is shown to developers as well and returns a
404. We do not want to show links that cannot be accessed by design.
Therefore, the link is only shown to users that are maintainers or above
for better user experience. The docs link (Learn more) was added since
it adds value for all.

Docs link: https://docs.gitlab.com/ee/ci/variables/#define-a-cicd-variable-in-the-ui
Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/476540
Changelog: fixed