Make merge trains more robust to ref deletion errors
Prior to this change, if ref deletion failed on the last retry of a
`MergeTrains::RefreshWorker` job, it could cause the train to get stuck
until a new job ran for the train. For example, say we have a multi-car
successful train where the first car's pipeline finished last. Under
normal circumstances, the `MergeTrains::RefreshWorker` job would merge
the associated MRs in order. But if the job is on its last retry, and an
error occurs in the post-merge ref cleanup for any of the cars, then the
merge sequence halts until the train gets refreshed again (such as when
a new car is added).

To fix this, we rescue and track ref cleanup errors. This is strictly
better than before, because the refs would have been orphaned either
way, but at least now the train does not get disrupted.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/496664

Changelog: fixed
EE: true