Add GraphQL API for Wiki notes
This adds GraphQL API backend for Wiki notes previously added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/163305,
complete with a WikiPageType, resolver, and all the glue code.

Additionally:

- Removed WikiPage::MetaPolicy and consolidated it into existing
WikiPagePolicy. There's no reason for the meta record to have a
separate set of permissions (at least for now).
- Fixed Note URL generation for group wikis and added a missing spec
for it. I can't easily separate this out of this MR, because some group
wiki specs rely on it working.

Changelog: added