Fix unlikely race on previous merge train car
It is technically possible for `car.first_car?` to return `false` but a
subsequent `car.prev` to return `nil`. When this occurs, we get a
`NoMethodError` from within the call to `mergeable_sha_and_message?`
below. I found at least one occurrence of this on the GitLab.com error
tracker (alas, the error is conflated with an unrelated one, making it
hard to find and link).

With this change, we eliminate this race by using `car.prev` for for both
checks.

No tests were added or changed because there's no sensible way to do so.
We would need to stub a call to `#first_car?` _that is no longer
present_.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/496664

Changelog: fixed
EE: true