Unmark project as pending delete if user lacks delete access
We mark the project as `pending_delete` during async project
deletion, signifying that deletion is in progress by the Sidekiq
worker. If the user loses access before deletion occurs, we should
unmark the project as deleted. This allows another user to restore the
project instead of the service simply returning and the project remains
in the database inaccessible to any user.

Changelog: fixed