Convert 20241022181600 to a schema migration
This migration caused the BBM to begin erroring on .com because
application code is updated _before_ post-deployment migrations run. For
.com the issue was addressed by pausing the migration until the
post-deployment migrations were able to run an delete it. This change
converts the migration to a schema migration so that the same issue does
not happen during self-managed upgrades.

Changelog: fixed