Fix GitLab Duo chat session error in the Web IDE
Fix GitLab Duo chat session error that occurs when a session concern
assumes that the Web IDE uses a session cookie to authenticate
HTTP requests to the GraphQL API, but the Web IDE is using an OAuth
session token instead.

In this context, getting the session.id property results in an error
because the underlying options property is not an instance of
ActionDispatch::Request::Session::Options, it is a hash with a key value
pair { skip: true }.

The fix checks if the options property is a hash and returns nil
in this scenario.

Changelog: fixed
EE: true