Use `deleted_at` column to mark groups in deletion process
The current 'pending_delete' column name is misleading. It suggests
future deletion, but actually indicates an ongoing deletion process.
The new column name better reflects the actual state of the group
and avoids confusion about the deletion status.

Here we don't need to bakfill this column because the only finder using
the current column i.e. `Organizations::GroupsFinder` is experimental.