Don't create approval rules without require_approval action
We skip approval rule creation for `any_merge_request` rules if
there is no `require_approval` action in the policy to reduce noice
with approvals.

However, we introduced `send_bot_action` and we didn't account for it
in the check.

This change changes that check to not create approval rules if the only
action is to send a bot message.

Changelog: fixed
EE: true