Create new WorkflowsInternal API for ai_workflows scope
This MR does not introduce any new API points but instead just moves
them from `Workflows` to `WorkflowsInternal`. The previous `Workflows`
API contains a mix of endpoints that are consumed by the Duo Workflow
Service as well as the VS Code extension in GitLab. The issue is that
the Duo Workflow Service only has an `ai_workflows` scope token which is
meant to be limited to only a narrow set of things it can do. The VS
Code extension has an `api` scoped token.

This MR splits the API up such that the endpoints needed by Duo Workflow
Service are all inside the `WorkflowsInternal` API.

Resolves https://gitlab.com/gitlab-org/gitlab/-/issues/500057

Changelog: other