Add admin role option for the default project creation protection
We have a `Default project creation protection` setting in
`Admin -> Settings -> General -> Visibility and access controls` that
controls the default setting for the group for the minimum access needed
to create a project in the group. We have a `No one` option that
prevents even the admins from creating projects in the group. Therefore,
we have added a new option `Administrators` for this setting, allowing
admins to create projects in the group by default.

With this new option selected, when admin mode is enabled, the admin
will need to enter admin mode to create new projects.

Changelog: added