Housekeeper do not push again if there are approvals
[GitLab Housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
is an automation tool we run in CI to create MRs that save developers
from repetitive tasks.

Something we've noticed a lot with Housekeeper is MRs like
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/170877 where
Housekeeper is running daily and it keeps pushing changes to the same
MR. Every time it pushes changes this resets all the approvals for the
MR. This seems to happen a lot where the maintainer has set to
auto-merge but we lose this because Housekeeper pushes before the
pipeline finishes. Even worse, since Housekeeper does not have
permission to trigger downstream pipelines, these pushes also trigger a
failing pipeline so it delays merging until the maintainer creates
another pipeline.

We discussed
[in Slack (internal)](https://gitlab.slack.com/archives/C06TH0AD26Q/p1730670849022289)
and it seems like the simplest workaround for now is to just not push
anymore once an MR has approvals. Under most circumstances if the MR is
already approved then we should leave it as-is. It's also trickier for
us to implement any more sophisticated ways to dermine if the MR should
be updated as keeps can generate a slightly different change each time
they run.

Since some people running Housekeeper manually might want to push anyway
(if they are still developing the keep and improving the MRs each time)
then we also needed a command line option to revert to the previous
behaviour.